refactor: update RadioGroup component
